requirements.txt
neo4j
openai
akshare
python-dotenv
mcp
colorama
transformers
torch